schema changes that happen before merge and hamber menu prep
